refactor(termiod): give the session's accounting its own modules
`session.rs` held five unrelated concerns in 2908 lines: the client backlog,
the VT sidecar's queue and vocabulary, foreground sampling, wire encoding, and
the session state machine itself. Only the last one depends on the others —
they are used *by* the session, not the reverse — which is what makes them
separable, rather than the line count.

Two are lifted here, the two that are pure policy:

- `session::backlog` — what a client may have in flight, and what happens when
  it stops keeping up. No PTY, no VT, no runtime; the decision to drop a
  straggler and resync it now lives in one place.
- `session::sidecar` — the sidecar's command and result vocabulary plus the
  budget that keeps a slow *parse* from costing unbounded memory, one hop
  upstream of the per-client one.

The compiler made the previously implicit boundaries explicit, which was the
point: every method the session reaches for had to be named `pub(crate)`
deliberately instead of being visible by accident of sharing a file. Two
constants stopped leaking as well — each module now answers `cap_description()`
rather than having four call sites restate its limit, so a number and the
sentence explaining it cannot drift apart.

The unit tests for both moved with them; they exercised the accounting, not the
session. One was reaching into a private atomic to ask whether the queue had
drained, which is now `is_drained()`.

Behaviour is unchanged — a pure move, plus visibility. 106 unit tests and every
integration suite pass. `session.rs` no longer imports an atomic type at all,
which is the check that the boundary is real.
